# app/marksman/internal/server/swagger/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /v1/datasource/{uid}/metric/{metric}:
        get:
            tags:
                - Datasource
            description: GetMetricLabelDetail returns one metric's labels and each label's values (detail view for a single metric).
            operationId: Datasource_GetMetricDetail
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
                - name: metric
                  in: path
                  description: Metric name to query label detail for (e.g. "http_requests_total", "up").
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/marksman.api.v1.MetricDetailItem'
...
    /v1/metric/strategy/{strategyUID}:
        get:
            tags:
                - StrategyMetric
            operationId: StrategyMetric_GetStrategyMetric
            parameters:
                - name: strategyUID
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/marksman.api.v1.StrategyMetricItem'
        post:
            tags:
                - StrategyMetric
            operationId: StrategyMetric_SaveStrategyMetric
            parameters:
                - name: strategyUID
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/marksman.api.v1.SaveStrategyMetricRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/marksman.api.v1.SaveStrategyMetricReply'
    /v1/metric/strategy/{strategyUID}/level:
        post:
            tags:
                - StrategyMetric
            operationId: StrategyMetric_SaveStrategyMetricLevel
            parameters:
                - name: strategyUID
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/marksman.api.v1.SaveStrategyMetricLevelRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/marksman.api.v1.SaveStrategyMetricLevelReply'
    /v1/metric/strategy/{strategyUID}/level/{levelUID}:
        get:
            tags:
                - StrategyMetric
            operationId: StrategyMetric_GetStrategyMetricLevel
            parameters:
                - name: strategyUID
                  in: path
                  required: true
                  schema:
                    type: string
                - name: levelUID
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/marksman.api.v1.StrategyMetricLevelItem'
        delete:
            tags:
                - StrategyMetric
            operationId: StrategyMetric_DeleteStrategyMetricLevel
            parameters:
                - name: strategyUID
                  in: path
                  required: true
                  schema:
                    type: string
                - name: levelUID
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/marksman.api.v1.DeleteStrategyMetricLevelReply'
    /v1/metric/strategy/{strategyUID}/level/{levelUID}/status:
        put:
            tags:
                - StrategyMetric
            operationId: StrategyMetric_UpdateStrategyMetricLevelStatus
            parameters:
                - name: strategyUID
                  in: path
                  required: true
                  schema:
                    type: string
                - name: levelUID
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/marksman.api.v1.UpdateStrategyMetricLevelStatusRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/marksman.api.v1.UpdateStrategyMetricLevelStatusReply'
...
    /v1/strategy:
        post:
            tags:
                - Strategy
            operationId: Strategy_CreateStrategy
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/marksman.api.v1.CreateStrategyRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/marksman.api.v1.CreateStrategyReply'
...
    /v1/strategy/{uid}:
        get:
            tags:
                - Strategy
            operationId: Strategy_GetStrategy
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/marksman.api.v1.StrategyItem'
        put:
            tags:
                - Strategy
            operationId: Strategy_UpdateStrategy
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/marksman.api.v1.UpdateStrategyRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/marksman.api.v1.UpdateStrategyReply'
        delete:
            tags:
                - Strategy
            operationId: Strategy_DeleteStrategy
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/marksman.api.v1.DeleteStrategyReply'
...
components:
    schemas:
...
        marksman.api.v1.CreateStrategyReply:
            type: object
            properties:
                uid:
                    type: string
        marksman.api.v1.CreateStrategyRequest:
            type: object
            properties:
                name:
                    type: string
                remark:
                    type: string
                type:
                    type: integer
                    format: enum
                driver:
                    type: integer
                    format: enum
                metadata:
                    type: object
                    additionalProperties:
                        type: string
                strategyGroupUID:
                    type: string
                status:
                    type: integer
                    format: enum
...
        marksman.api.v1.DeleteStrategyMetricLevelReply:
            type: object
            properties: {}
        marksman.api.v1.DeleteStrategyReply:
            type: object
            properties: {}
...
        marksman.api.v1.LevelItem:
            type: object
            properties:
                uid:
                    type: string
                name:
                    type: string
                remark:
                    type: string
                status:
                    type: integer
                    format: enum
                metadata:
                    type: object
                    additionalProperties:
                        type: string
                createdAt:
                    type: string
                updatedAt:
                    type: string
                bgColor:
                    type: string
                type:
                    type: integer
                    format: enum
...
        marksman.api.v1.MetricDetailItem:
            type: object
            properties:
                name:
                    type: string
                help:
                    type: string
                unit:
                    type: string
                type:
                    type: string
                labels:
                    type: array
                    items:
                        $ref: '#/components/schemas/marksman.api.v1.MetricLabelItem'
        marksman.api.v1.MetricLabelItem:
            type: object
            properties:
                name:
                    type: string
                values:
                    type: array
                    items:
                        type: string
...
        marksman.api.v1.SaveStrategyMetricLevelReply:
            type: object
            properties: {}
        marksman.api.v1.SaveStrategyMetricLevelRequest:
            type: object
            properties:
                strategyUID:
                    type: string
                levelUID:
                    type: string
                mode:
                    type: integer
                    format: enum
                condition:
                    type: integer
                    format: enum
                values:
                    type: array
                    items:
                        type: number
                        format: double
                duration:
                    pattern: ^-?(?:0|[1-9][0-9]{0,11})(?:\.[0-9]{1,9})?s$
                    type: string
                status:
                    type: integer
                    format: enum
        marksman.api.v1.SaveStrategyMetricReply:
            type: object
            properties: {}
        marksman.api.v1.SaveStrategyMetricRequest:
            type: object
            properties:
                strategyUID:
                    type: string
                expr:
                    type: string
                labels:
                    type: object
                    additionalProperties:
                        type: string
                datasourceUIDs:
                    type: array
                    items:
                        type: string
                summary:
                    type: string
                description:
                    type: string
...
        marksman.api.v1.StrategyGroupItem:
            type: object
            properties:
                uid:
                    type: string
                name:
                    type: string
                remark:
                    type: string
                status:
                    type: integer
                    format: enum
                metadata:
                    type: object
                    additionalProperties:
                        type: string
                createdAt:
                    type: string
                updatedAt:
                    type: string
...
        marksman.api.v1.StrategyItem:
            type: object
            properties:
                uid:
                    type: string
                name:
                    type: string
                remark:
                    type: string
                type:
                    type: integer
                    format: enum
                driver:
                    type: integer
                    format: enum
                status:
                    type: integer
                    format: enum
                metadata:
                    type: object
                    additionalProperties:
                        type: string
                createdAt:
                    type: string
                updatedAt:
                    type: string
                strategyGroup:
                    $ref: '#/components/schemas/marksman.api.v1.StrategyGroupItem'
...
        marksman.api.v1.StrategyMetricItem:
            type: object
            properties:
                strategyUID:
                    type: string
                expr:
                    type: string
                labels:
                    type: object
                    additionalProperties:
                        type: string
                summary:
                    type: string
                description:
                    type: string
                datasourceUIDs:
                    type: array
                    items:
                        type: string
                levels:
                    type: array
                    items:
                        $ref: '#/components/schemas/marksman.api.v1.StrategyMetricLevelItem'
                createdAt:
                    type: string
                updatedAt:
                    type: string
                strategy:
                    $ref: '#/components/schemas/marksman.api.v1.StrategyItem'
        marksman.api.v1.StrategyMetricLevelItem:
            type: object
            properties:
                levelUID:
                    type: string
                strategyUID:
                    type: string
                level:
                    $ref: '#/components/schemas/marksman.api.v1.LevelItem'
                mode:
                    type: integer
                    format: enum
                condition:
                    type: integer
                    format: enum
                values:
                    type: array
                    items:
                        type: number
                        format: double
                duration:
                    pattern: ^-?(?:0|[1-9][0-9]{0,11})(?:\.[0-9]{1,9})?s$
                    type: string
                status:
                    type: integer
                    format: enum
...
        marksman.api.v1.UpdateStrategyMetricLevelStatusReply:
            type: object
            properties: {}
        marksman.api.v1.UpdateStrategyMetricLevelStatusRequest:
            type: object
            properties:
                strategyUID:
                    type: string
                levelUID:
                    type: string
                status:
                    type: integer
                    format: enum
        marksman.api.v1.UpdateStrategyReply:
            type: object
            properties: {}
        marksman.api.v1.UpdateStrategyRequest:
            type: object
            properties:
                uid:
                    type: string
                name:
                    type: string
                remark:
                    type: string
                metadata:
                    type: object
                    additionalProperties:
                        type: string
                strategyGroupUID:
                    type: string
                type:
                    type: integer
                    format: enum
                driver:
                    type: integer
                    format: enum
...