# app/marksman/pkg/api/v1/strategy_metric_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v5.29.2
// source: marksman/api/v1/strategy_metric.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationStrategyMetricDeleteStrategyMetricLevel = "/marksman.api.v1.StrategyMetric/DeleteStrategyMetricLevel"
const OperationStrategyMetricGetStrategyMetric = "/marksman.api.v1.StrategyMetric/GetStrategyMetric"
const OperationStrategyMetricGetStrategyMetricLevel = "/marksman.api.v1.StrategyMetric/GetStrategyMetricLevel"
const OperationStrategyMetricSaveStrategyMetric = "/marksman.api.v1.StrategyMetric/SaveStrategyMetric"
const OperationStrategyMetricSaveStrategyMetricLevel = "/marksman.api.v1.StrategyMetric/SaveStrategyMetricLevel"
const OperationStrategyMetricUpdateStrategyMetricLevelStatus = "/marksman.api.v1.StrategyMetric/UpdateStrategyMetricLevelStatus"

type StrategyMetricHTTPServer interface {
	DeleteStrategyMetricLevel(context.Context, *DeleteStrategyMetricLevelRequest) (*DeleteStrategyMetricLevelReply, error)
	GetStrategyMetric(context.Context, *GetStrategyMetricRequest) (*StrategyMetricItem, error)
	GetStrategyMetricLevel(context.Context, *GetStrategyMetricLevelRequest) (*StrategyMetricLevelItem, error)
	SaveStrategyMetric(context.Context, *SaveStrategyMetricRequest) (*SaveStrategyMetricReply, error)
	SaveStrategyMetricLevel(context.Context, *SaveStrategyMetricLevelRequest) (*SaveStrategyMetricLevelReply, error)
	UpdateStrategyMetricLevelStatus(context.Context, *UpdateStrategyMetricLevelStatusRequest) (*UpdateStrategyMetricLevelStatusReply, error)
}

func RegisterStrategyMetricHTTPServer(s *http.Server, srv StrategyMetricHTTPServer) {
	r := s.Route("/")
	r.POST("/v1/metric/strategy/{strategyUID}", _StrategyMetric_SaveStrategyMetric0_HTTP_Handler(srv))
	r.GET("/v1/metric/strategy/{strategyUID}", _StrategyMetric_GetStrategyMetric0_HTTP_Handler(srv))
	r.POST("/v1/metric/strategy/{strategyUID}/level", _StrategyMetric_SaveStrategyMetricLevel0_HTTP_Handler(srv))
	r.PUT("/v1/metric/strategy/{strategyUID}/level/{levelUID}/status", _StrategyMetric_UpdateStrategyMetricLevelStatus0_HTTP_Handler(srv))
	r.DELETE("/v1/metric/strategy/{strategyUID}/level/{levelUID}", _StrategyMetric_DeleteStrategyMetricLevel0_HTTP_Handler(srv))
	r.GET("/v1/metric/strategy/{strategyUID}/level/{levelUID}", _StrategyMetric_GetStrategyMetricLevel0_HTTP_Handler(srv))
}

func _StrategyMetric_SaveStrategyMetric0_HTTP_Handler(srv StrategyMetricHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in SaveStrategyMetricRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationStrategyMetricSaveStrategyMetric)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.SaveStrategyMetric(ctx, req.(*SaveStrategyMetricRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*SaveStrategyMetricReply)
		return ctx.Result(200, reply)
	}
}

func _StrategyMetric_GetStrategyMetric0_HTTP_Handler(srv StrategyMetricHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetStrategyMetricRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationStrategyMetricGetStrategyMetric)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetStrategyMetric(ctx, req.(*GetStrategyMetricRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*StrategyMetricItem)
		return ctx.Result(200, reply)
	}
}

func _StrategyMetric_SaveStrategyMetricLevel0_HTTP_Handler(srv StrategyMetricHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in SaveStrategyMetricLevelRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationStrategyMetricSaveStrategyMetricLevel)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.SaveStrategyMetricLevel(ctx, req.(*SaveStrategyMetricLevelRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*SaveStrategyMetricLevelReply)
		return ctx.Result(200, reply)
	}
}

func _StrategyMetric_UpdateStrategyMetricLevelStatus0_HTTP_Handler(srv StrategyMetricHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateStrategyMetricLevelStatusRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationStrategyMetricUpdateStrategyMetricLevelStatus)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateStrategyMetricLevelStatus(ctx, req.(*UpdateStrategyMetricLevelStatusRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateStrategyMetricLevelStatusReply)
		return ctx.Result(200, reply)
	}
}

func _StrategyMetric_DeleteStrategyMetricLevel0_HTTP_Handler(srv StrategyMetricHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DeleteStrategyMetricLevelRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationStrategyMetricDeleteStrategyMetricLevel)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DeleteStrategyMetricLevel(ctx, req.(*DeleteStrategyMetricLevelRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*DeleteStrategyMetricLevelReply)
		return ctx.Result(200, reply)
	}
}

func _StrategyMetric_GetStrategyMetricLevel0_HTTP_Handler(srv StrategyMetricHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetStrategyMetricLevelRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationStrategyMetricGetStrategyMetricLevel)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetStrategyMetricLevel(ctx, req.(*GetStrategyMetricLevelRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*StrategyMetricLevelItem)
		return ctx.Result(200, reply)
	}
}

type StrategyMetricHTTPClient interface {
	DeleteStrategyMetricLevel(ctx context.Context, req *DeleteStrategyMetricLevelRequest, opts ...http.CallOption) (rsp *DeleteStrategyMetricLevelReply, err error)
	GetStrategyMetric(ctx context.Context, req *GetStrategyMetricRequest, opts ...http.CallOption) (rsp *StrategyMetricItem, err error)
	GetStrategyMetricLevel(ctx context.Context, req *GetStrategyMetricLevelRequest, opts ...http.CallOption) (rsp *StrategyMetricLevelItem, err error)
	SaveStrategyMetric(ctx context.Context, req *SaveStrategyMetricRequest, opts ...http.CallOption) (rsp *SaveStrategyMetricReply, err error)
	SaveStrategyMetricLevel(ctx context.Context, req *SaveStrategyMetricLevelRequest, opts ...http.CallOption) (rsp *SaveStrategyMetricLevelReply, err error)
	UpdateStrategyMetricLevelStatus(ctx context.Context, req *UpdateStrategyMetricLevelStatusRequest, opts ...http.CallOption) (rsp *UpdateStrategyMetricLevelStatusReply, err error)
}

type StrategyMetricHTTPClientImpl struct {
	cc *http.Client
}

func NewStrategyMetricHTTPClient(client *http.Client) StrategyMetricHTTPClient {
	return &StrategyMetricHTTPClientImpl{client}
}

func (c *StrategyMetricHTTPClientImpl) DeleteStrategyMetricLevel(ctx context.Context, in *DeleteStrategyMetricLevelRequest, opts ...http.CallOption) (*DeleteStrategyMetricLevelReply, error) {
	var out DeleteStrategyMetricLevelReply
	pattern := "/v1/metric/strategy/{strategyUID}/level/{levelUID}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationStrategyMetricDeleteStrategyMetricLevel))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *StrategyMetricHTTPClientImpl) GetStrategyMetric(ctx context.Context, in *GetStrategyMetricRequest, opts ...http.CallOption) (*StrategyMetricItem, error) {
	var out StrategyMetricItem
	pattern := "/v1/metric/strategy/{strategyUID}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationStrategyMetricGetStrategyMetric))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *StrategyMetricHTTPClientImpl) GetStrategyMetricLevel(ctx context.Context, in *GetStrategyMetricLevelRequest, opts ...http.CallOption) (*StrategyMetricLevelItem, error) {
	var out StrategyMetricLevelItem
	pattern := "/v1/metric/strategy/{strategyUID}/level/{levelUID}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationStrategyMetricGetStrategyMetricLevel))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *StrategyMetricHTTPClientImpl) SaveStrategyMetric(ctx context.Context, in *SaveStrategyMetricRequest, opts ...http.CallOption) (*SaveStrategyMetricReply, error) {
	var out SaveStrategyMetricReply
	pattern := "/v1/metric/strategy/{strategyUID}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationStrategyMetricSaveStrategyMetric))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *StrategyMetricHTTPClientImpl) SaveStrategyMetricLevel(ctx context.Context, in *SaveStrategyMetricLevelRequest, opts ...http.CallOption) (*SaveStrategyMetricLevelReply, error) {
	var out SaveStrategyMetricLevelReply
	pattern := "/v1/metric/strategy/{strategyUID}/level"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationStrategyMetricSaveStrategyMetricLevel))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *StrategyMetricHTTPClientImpl) UpdateStrategyMetricLevelStatus(ctx context.Context, in *UpdateStrategyMetricLevelStatusRequest, opts ...http.CallOption) (*UpdateStrategyMetricLevelStatusReply, error) {
	var out UpdateStrategyMetricLevelStatusReply
	pattern := "/v1/metric/strategy/{strategyUID}/level/{levelUID}/status"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationStrategyMetricUpdateStrategyMetricLevelStatus))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}